refactor: componentization of widgets
