+++ lib/shared/providers.dart
@@ -1,0 +1,3 @@
+import 'package:hooks_riverpod/hooks_riverpod.dart';
+
+final visibleProvider = StateProvider((ref) => true);
