+++ pubspec.yaml
@@ -37,6 +37,8 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   intl: ^0.17.0
+  decimal: ^2.3.0
+  hooks_riverpod: ^1.0.4
 
 dev_dependencies:
   flutter_test:
